fix: cast bfloat16 FPN embeddings to float32 before numpy conversion
Adds .float() before .numpy() on SAM3 vision encoder FPN outputs to prevent
'RuntimeError: Got unsupported ScalarType BFloat16' when converting to numpy:

- Add .float() cast in complete_reid.py batch embedding extraction
- Add .float() cast in initial_seeding_video.py crop embedding extraction
- Add .float() cast in seeding_common.py batch frame embedding extraction

+++ label_studio_ml/examples/segment_anything_3_video/complete_reid.py
@@ -731,7 +731,7 @@ def _compute_track_features_sam3(
             # fpn_hidden_states[0]: highest-res FPN feature (B, C, H, W)
             feat = vision_output.fpn_hidden_states[0]
             if hasattr(feat, "detach"):
-                emb_batch = feat.mean(dim=(2, 3)).detach().cpu().numpy()
+                emb_batch = feat.mean(dim=(2, 3)).detach().cpu().float().numpy()
             else:
                 emb_batch = feat.mean(axis=(2, 3))
